requirements.txt
pyyaml
loguru
cx-freeze
opencv-python